requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
transformers==4.35.2
torch==2.1.1
Flask==3.0.0
--extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.17.1+cpu
torchaudio==2.2.1+cpu
scikit-learn==1.3.2 